# Conflicts:
#	ansible/project/files/brook-template.service
#	ansible/project/files/clean.sh
#	ansible/project/files/gost@.service
#	ansible/project/files/iptables.sh

+++ ansible/project/files/iptables.sh
@@ -2,11 +2,6 @@
 
 SUDO=$(if [ $(id -u $whoami) -gt 0 ]; then echo "sudo "; fi)
 [ -z $SUDO ] || sudo -n true 2>/dev/null || (echo "Failed to use sudo" && exit 1)
-IFACE=$(ip route show | grep default | grep -Po '(?<=dev )(\w+)')
-INET=$(ip address show $IFACE scope global |  awk '/inet / {split($2,var,"/"); print var[1]}')
-# Only support one ip now
-INET=$(echo $INET | awk '{print $1}' | grep -Po "^((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$")
-[ -z $INET ] && echo "No valid interface ipv4 addresses found" && exit 1
 TYPE="ALL"
 LOCAL_PORT=0
 REMOTE_IP=0
@@ -14,32 +9,48 @@ REMOTE_PORT=0
 
 check_system () {
     source '/etc/os-release'
-    if [[ $ID = "centos" ]]; then
+    if [[ $ID == "centos" ]]; then
         OS_FAMILY="centos"
         UPDATE="$SUDO yum makecache -y"
-        INSTALL="$SUDO yum install -y iptables"
-    elif [[ $ID = "debian" || $ID = "ubuntu" ]]; then
+        INSTALL="$SUDO yum install -y"
+    elif [[ $ID == "debian" || $ID == "ubuntu" ]]; then
         OS_FAMILY="debian"
         UPDATE="$SUDO apt update -y"
-        INSTALL="$SUDO apt install -y iptables"
-    else
-        echo -e "System $ID ${VERSION_ID} is not supported now"
-        exit 1
+        INSTALL="$SUDO apt install -y --no-install-recommends"
+    elif [[ $ID == "alpine" ]]; then
+        OS_FAMILY="alpine"
+        UPDATE="$SUDO apk update"
+        INSTALL="$SUDO apk add --no-cache"
     fi
+    # Not force to exit if the system is not supported
+    systemctl --version > /dev/null 2>&1 && IS_SYSTEMD=1
 }
 
-install_ipt () {
-    iptables -V > /dev/null || $INSTALL || ($UPDATE && $INSTALL)
-    if [ $? -ne 0 ]; then
-        echo -e "Failed to install iptables"
-        exit 1
+get_ips () {
+    IFACE=$(ip route show | grep default | awk -F 'dev ' '{ print $2; }' | awk '{ print $1; }')
+    INET=$(ip address show $IFACE scope global |  awk '/inet / {split($2,var,"/"); print var[1]}')
+    INET=$(echo $INET | xargs -n 1 | grep -Eo "^((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$" | sort -u)
+    [ -z $INET ] && echo "No valid interface ipv4 addresses found" && exit 1
+}
+
+install_deps () {
+    iptables -V > /dev/null || $INSTALL iptables || ($UPDATE && $INSTALL iptables) || (echo "Failed to install iptables" && exit 1)
+    ip a > /dev/null && return 0
+    if [[ $OS_FAMILY == "centos" ]]; then
+        $INSTALL iproute || ($UPDATE && $INSTALL iproute) || (echo "Failed to install iproute" && exit 1)
+    elif [[ $OS_FAMILY == "debian" ]]; then
+        $INSTALL iproute2 || ($UPDATE && $INSTALL iproute2) || (echo "Failed to install iproute2" && exit 1)
+    else
+        echo "ip command not found" && exit 1
     fi
 }
 
 delete_service () {
     [[ -z $1 ]] && SERVICE="aurora@${LOCAL_PORT}.service" || SERVICE=$1
-    systemctl is-active --quiet $SERVICE > /dev/null 2>&1 && $SUDO systemctl stop $SERVICE
-    systemctl is-enabled --quiet $SERVICE > /dev/null 2>&1 && $SUDO systemctl disable $SERVICE
+    if [[ $IS_SYSTEMD -eq 1 ]]; then
+        systemctl is-active --quiet $SERVICE > /dev/null 2>&1 && $SUDO systemctl stop $SERVICE
+        systemctl is-enabled --quiet $SERVICE > /dev/null 2>&1 && $SUDO systemctl disable $SERVICE
+    fi
 }
 
 disable_firewall () {
@@ -50,31 +61,31 @@ disable_firewall () {
 }
 
 check_ipt_restore_file () {
-    if [ $OS_FAMILY = "centos" ]; then
+    if [[ $OS_FAMILY == "centos" ]]; then
         IPT_PATH="/etc/sysconfig"
-    else
+    elif [[ $OS_FAMILY == "debian" ]]; then
         IPT_PATH="/etc/iptables"
     fi
-    if [ ! -d $IPT_PATH ]; then
+    if [[ -n $IPT_PATH && ! -d $IPT_PATH ]]; then
         $SUDO mkdir -p $IPT_PATH
     fi
-    if [ $OS_FAMILY = "centos" ]; then
+    if [[ $OS_FAMILY == "centos" ]]; then
         IPT_RESTORE_FILE=$IPT_PATH/iptables
-    else
+    elif [[ $OS_FAMILY == "debian" ]]; then
         IPT_RESTORE_FILE=$IPT_PATH/rules.v4
     fi
-    if [ ! -f $IPT_RESTORE_FILE ]; then
+    if [[ -n $IPT_RESTORE_FILE && ! -f $IPT_RESTORE_FILE ]]; then
         $SUDO touch $IPT_RESTORE_FILE
     fi
 }
 
 install_ipt_service () {
-    if [ $OS_FAMILY = "centos" ]; then
+    if [[ $OS_FAMILY == "centos" ]]; then
         RULE_PATH="/etc/sysconfig/iptables"
-    elif [ $OS_FAMILY = "debian" ]; then
+    elif [[ $OS_FAMILY == "debian" ]]; then
         RULE_PATH="/etc/iptables/rules.v4"
     fi
-    $SUDO cat > /etc/systemd/system/iptables-restore.service <<EOF
+    [[ ! -z $RULE_PATH ]] && $SUDO cat > /etc/systemd/system/iptables-restore.service <<EOF
 [Unit]
 Description=Restore iptables rule by Aurora Admin Panel
 
@@ -115,45 +126,54 @@ $SUDO systemctl start iptables-outipcheck.timer
 }
 
 check_ipt_service () {
-    # add out ip changed check timer for iptable port forward rules
-    ! systemctl is-enabled --quiet iptables-outipcheck.timer > /dev/null 2>&1 && install_ipt_out_ip_check_service
-    # service to support auto save iptable forward rules
-    ! systemctl is-enabled --quiet iptables-restore.service > /dev/null 2>&1 && install_ipt_service \
-    $SUDO systemctl daemon-reload && \
-    $SUDO systemctl enable iptables-restore.service
-    ! systemctl is-enabled --quiet iptables-restore.service && \
-    echo -e "Failed to install iptables restore service" && exit 1
+    if [[ $IS_SYSTEMD -eq 1 ]]; then
+        # add out ip changed check timer for iptable port forward rules
+        ! systemctl is-enabled --quiet iptables-outipcheck.timer > /dev/null 2>&1 && install_ipt_out_ip_check_service
+        ! systemctl is-enabled --quiet iptables-restore.service > /dev/null 2>&1 && install_ipt_service \
+        $SUDO systemctl daemon-reload && \
+        $SUDO systemctl enable iptables-restore.service
+        ! systemctl is-enabled --quiet iptables-restore.service && \
+        echo "Failed to install iptables restore service"
+    fi
+    # Not force to exit if the system does not use the systemd
 }
 
 save_iptables () {
     check_ipt_restore_file
-    $SUDO iptables-save -c | $SUDO tee $IPT_RESTORE_FILE > /dev/null
+    [[ -f $IPT_RESTORE_FILE ]] && $SUDO iptables-save -c | $SUDO tee $IPT_RESTORE_FILE > /dev/null
 }
 
 set_forward () {
+    [[ $(cat /proc/sys/net/ipv4/ip_forward) -eq 1 ]] && return 0
     if [[ -z $($SUDO cat /etc/sysctl.conf | grep "net.ipv4.ip_forward") ]]; then
         echo "net.ipv4.ip_forward = 1" | $SUDO tee -a /etc/sysctl.conf > /dev/null
-    elif [[ -n $($SUDO cat /etc/sysctl.conf | grep "net.ipv4.ip_forward" | grep "0") ]]; then
+    else
         sed -i "s/.*net.ipv4.ip_forward.*/net.ipv4.ip_forward = 1/g" /etc/sysctl.conf > /dev/null
     fi
     $SUDO sysctl -p > /dev/null
+    # check and make sure ip_forward enabled
+    [[ $(cat /proc/sys/net/ipv4/ip_forward) -ne 1 ]] && echo "Cannot enable ipv4 forward for iptables" && exit 1
 }
 
 forward () {
-    set_forward
+    set_forward || exit 1
     if [ $TYPE == "ALL" ] || [ $TYPE == "TCP" ]
     then
-        $SUDO iptables -t nat -A PREROUTING -p tcp --dport $LOCAL_PORT -j DNAT --to-destination $REMOTE_IP:$REMOTE_PORT  -m comment --comment "FORWARD $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT" && \
-        $SUDO iptables -t nat -A POSTROUTING -d $REMOTE_IP -p tcp --dport $REMOTE_PORT -j SNAT --to-source $INET -m comment --comment "BACKWARD $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT" && \
-        $SUDO iptables -I FORWARD -p tcp -d $REMOTE_IP --dport $REMOTE_PORT -j ACCEPT -m comment --comment "UPLOAD $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT" && \
-        $SUDO iptables -I FORWARD -p tcp -s $REMOTE_IP -j ACCEPT -m comment --comment "DOWNLOAD $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT" || (echo "Failed to create tcp forward rules" && exit 1)
+        for SNATIP in $INET; do
+            $SUDO iptables -t nat -A POSTROUTING -d $REMOTE_IP -p tcp --dport $REMOTE_PORT -j SNAT --to-source $SNATIP -m comment --comment "BACKWARD $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT"
+        done
+        $SUDO iptables -t nat -A PREROUTING -p tcp --dport $LOCAL_PORT -j DNAT --to-destination $REMOTE_IP:$REMOTE_PORT  -m comment --comment "FORWARD $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT"
+        $SUDO iptables -I FORWARD -p tcp -d $REMOTE_IP --dport $REMOTE_PORT -j ACCEPT -m comment --comment "UPLOAD $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT"
+        $SUDO iptables -I FORWARD -p tcp -s $REMOTE_IP -j ACCEPT -m comment --comment "DOWNLOAD $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT"
     fi
     if [ $TYPE == "ALL" ] || [ $TYPE == "UDP" ]
     then
-        $SUDO iptables -t nat -A PREROUTING -p udp --dport $LOCAL_PORT -j DNAT --to-destination $REMOTE_IP:$REMOTE_PORT  -m comment --comment "FORWARD $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT" && \
-        $SUDO iptables -t nat -A POSTROUTING -d $REMOTE_IP -p udp --dport $REMOTE_PORT -j SNAT --to-source $INET -m comment --comment "BACKWARD $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT" && \
-        $SUDO iptables -I FORWARD -p udp -d $REMOTE_IP --dport $REMOTE_PORT -j ACCEPT -m comment --comment "UPLOAD-UDP $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT" && \
-        $SUDO iptables -I FORWARD -p udp -s $REMOTE_IP -j ACCEPT -m comment --comment "DOWNLOAD-UDP $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT" || (echo "Failed to create udp forward rules" && exit 1)
+        for SNATIP in $INET; do
+            $SUDO iptables -t nat -A POSTROUTING -d $REMOTE_IP -p udp --dport $REMOTE_PORT -j SNAT --to-source $SNATIP -m comment --comment "BACKWARD $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT"
+        done
+        $SUDO iptables -t nat -A PREROUTING -p udp --dport $LOCAL_PORT -j DNAT --to-destination $REMOTE_IP:$REMOTE_PORT  -m comment --comment "FORWARD $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT"
+        $SUDO iptables -I FORWARD -p udp -d $REMOTE_IP --dport $REMOTE_PORT -j ACCEPT -m comment --comment "UPLOAD-UDP $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT"
+        $SUDO iptables -I FORWARD -p udp -s $REMOTE_IP -j ACCEPT -m comment --comment "DOWNLOAD-UDP $LOCAL_PORT->$REMOTE_IP:$REMOTE_PORT"
     fi
     save_iptables
 }
@@ -257,7 +277,7 @@ fi
 [ -n $2 ] && LOCAL_PORT=$2
 [ "$OPERATION" != "outipcheck" ] && [ -z $LOCAL_PORT ] && echo "Illegal port $PORT" && exit 1
 [ -n $3 ] && REMOTE_IP=$3
-REMOTE_IP=$(echo $REMOTE_IP | grep -Po "^((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$")
+REMOTE_IP=$(echo $REMOTE_IP | grep -Eo "^((25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$")
 if [ $OPERATION == "forward" ] && [ -z $REMOTE_IP ]
 then
     echo "Unknow remote ip for operation $OPERATION" && exit 1
@@ -269,9 +289,10 @@ then
 fi
 
 check_system
-install_ipt
+install_deps
 disable_firewall
 check_ipt_service
+get_ips
 if [ $OPERATION == "forward" ]; then
     list
     delete_service
@@ -285,6 +306,8 @@ elif [ $OPERATION == "list" ]; then
     list
 elif [ $OPERATION == "list_all" ]; then
     list_all
+elif [ $OPERATION == "delete_service" ]; then
+    delete_service
 elif [ $OPERATION == "delete" ]; then
     delete
 elif [ $OPERATION == "reset" ]; then
